requirements.txt
apscheduler
requests
hmac
hashlib
secure-smtplib
PyYAML